backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["AlexanderFarkas <sashafarkas322@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.2"}
sqlalchemy = "^2.0.35"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
python-dotenv = "^1.0.1"
argon2-cffi = "^23.1.0"
psycopg = {extras = ["binary", "pool"], version = "^3.2.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
